Refactoring for tests
