+++ tests/PSRule.Tests/TestObject.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace PSRule.Pipeline;
+
+internal sealed class TestObject
+{
+    public string Name { get; set; }
+}
